Manually compile libargon2 instead of using make
As the libargon2 Makefile shells out to `uname -s` and `uname -m`, it
passes invalid linker flags when cross-compiling (as we're compiling for
one OS but on a host machine that might use another).

Instead of bending over backwards to work around this, abandon using
their Makefile altogether and instead reproduce the specific logic we
need, i.e.

1. Compile each of the core object files in a position independent way
2. Create a static archive of those object files

This way, we only build what we need and hopefully in a way that is a
static library on as many platforms as possible.