Support versionless hashes on JRuby
While hashing a password with `Argon2id::Password.create` will always
return an encoded hash with version 0x13 (19), we should support
verifying hashes of version 0x10 (16) as well even when they don't
specify the version in the hash. This already works for platforms using
the C extension but we need to explicitly pass the version when
generating the output on JRuby.