Specify Argon2 version on JRuby
While the default version of Argon2's algorithm is currently version 13
(represented by the decimal 19 in the encoded hash), be explicit on
JRuby in case jruby-openssl ever changes so that it is consistent with
the C version.

While we're at it, re-order the parameters to match the order given to
the C API (though this doesn't actually match the order presented in the
encoded hash where memory cost precedes time cost).